fb: fix recents animation on new immersive mode
thanks here as well for identify the problem to Kiril.
But I changed the implementation. Beside that we give the user
the ability to setup different navbar heights IMO there is absolutly
no need to use an observer to get the status. We can fetch this
from WindowManagerService to get all needed information.

So at least this patch fetches needed values from WindowManager
and calculates in the missing heights

Change-Id: I3ae57ba0a575624f8aea75ce8d48041f5a1c1c42

Conflicts:
	core/java/android/view/WindowManagerPolicy.java